r733: change dungeon-role being shown everywhere (fix #5 )
A new option was added to the dungeon-role status where the player can choose to show the icon everywhere or only when in a dungeon.
If the options is disabled, role icons are only shown when in a party group and inside a dungeon.
If the option is enabled, role icons are shown everywhere, whether solo or in a group, all depends on group members specs.